Add PartitionLoader to load partitions
Add DelegatingPartitionLoader to delegate the partion
loading to ManifestPartitionLoader when Partition retrieved
from Metastore has list of fileNames, sizes in it.
