+++ presto-hive/src/main/java/com/facebook/presto/hive/BackgroundHiveSplitLoader.java
@@ -14,112 +14,39 @@
 package com.facebook.presto.hive;
 
 import com.facebook.presto.common.predicate.Domain;
-import com.facebook.presto.hive.HiveBucketing.HiveBucketFilter;
-import com.facebook.presto.hive.HiveSplit.BucketConversion;
-import com.facebook.presto.hive.filesystem.ExtendedFileSystem;
-import com.facebook.presto.hive.metastore.Column;
-import com.facebook.presto.hive.metastore.Partition;
-import com.facebook.presto.hive.metastore.Storage;
+import com.facebook.presto.hive.StoragePartitionLoader.BucketSplitInfo;
 import com.facebook.presto.hive.metastore.Table;
-import com.facebook.presto.hive.util.HiveFileIterator.NestedDirectoryNotAllowedException;
-import com.facebook.presto.hive.util.InternalHiveSplitFactory;
 import com.facebook.presto.hive.util.ResumableTask;
 import com.facebook.presto.hive.util.ResumableTasks;
 import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.PrestoException;
-import com.facebook.presto.spi.SchemaTableName;
-import com.google.common.base.Suppliers;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterators;
-import com.google.common.io.CharStreams;
 import com.google.common.util.concurrent.ListenableFuture;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.FileStatus;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.PathFilter;
-import org.apache.hadoop.hive.ql.io.SymlinkTextInputFormat;
-import org.apache.hadoop.mapred.FileInputFormat;
-import org.apache.hadoop.mapred.FileSplit;
-import org.apache.hadoop.mapred.InputFormat;
-import org.apache.hadoop.mapred.InputSplit;
-import org.apache.hadoop.mapred.JobConf;
-import org.apache.hadoop.mapred.TextInputFormat;
-import org.apache.hudi.hadoop.HoodieParquetInputFormat;
-import org.apache.hudi.hadoop.HoodieROTablePathFilter;
-import org.apache.hudi.hadoop.realtime.HoodieParquetRealtimeInputFormat;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.lang.annotation.Annotation;
-import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Deque;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Optional;
-import java.util.Properties;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.Executor;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.function.IntPredicate;
-import java.util.function.Supplier;
 
-import static com.facebook.presto.hive.HiveBucketing.getVirtualBucketNumber;
-import static com.facebook.presto.hive.HiveColumnHandle.pathColumnHandle;
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_BAD_DATA;
 import static com.facebook.presto.hive.HiveErrorCode.HIVE_FILESYSTEM_ERROR;
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_INVALID_BUCKET_FILES;
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_INVALID_METADATA;
-import static com.facebook.presto.hive.HiveErrorCode.HIVE_INVALID_PARTITION_VALUE;
 import static com.facebook.presto.hive.HiveErrorCode.HIVE_UNKNOWN_ERROR;
-import static com.facebook.presto.hive.HiveSessionProperties.getMaxInitialSplitSize;
-import static com.facebook.presto.hive.HiveSessionProperties.getNodeSelectionStrategy;
-import static com.facebook.presto.hive.HiveSessionProperties.isUseListDirectoryCache;
-import static com.facebook.presto.hive.HiveUtil.getFooterCount;
-import static com.facebook.presto.hive.HiveUtil.getHeaderCount;
-import static com.facebook.presto.hive.HiveUtil.getInputFormat;
-import static com.facebook.presto.hive.NestedDirectoryPolicy.FAIL;
-import static com.facebook.presto.hive.NestedDirectoryPolicy.IGNORED;
-import static com.facebook.presto.hive.NestedDirectoryPolicy.RECURSE;
-import static com.facebook.presto.hive.S3SelectPushdown.shouldEnablePushdownForTable;
-import static com.facebook.presto.hive.metastore.MetastoreUtil.checkCondition;
-import static com.facebook.presto.hive.metastore.MetastoreUtil.getHiveSchema;
-import static com.facebook.presto.hive.metastore.MetastoreUtil.getPartitionLocation;
-import static com.facebook.presto.hive.util.ConfigurationUtils.toJobConf;
-import static com.facebook.presto.spi.StandardErrorCode.NOT_SUPPORTED;
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.collect.ImmutableList.toImmutableList;
-import static com.google.common.collect.Streams.stream;
 import static com.google.common.util.concurrent.Futures.immediateFuture;
-import static java.lang.Math.max;
-import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
-import static org.apache.hadoop.hive.common.FileUtils.HIDDEN_FILES_PATH_FILTER;
 
 public class BackgroundHiveSplitLoader
         implements HiveSplitLoader
 {
     private static final ListenableFuture<?> COMPLETED_FUTURE = immediateFuture(null);
 
-    private final Table table;
-    private final Optional<Domain> pathDomain;
-    private final Optional<BucketSplitInfo> tableBucketInfo;
-    private final HdfsEnvironment hdfsEnvironment;
-    private final HdfsContext hdfsContext;
-    private final NamenodeStats namenodeStats;
-    private final DirectoryLister directoryLister;
     private final int loaderConcurrency;
-    private final boolean recursiveDirWalkerEnabled;
     private final Executor executor;
-    private final ConnectorSession session;
     private final ConcurrentLazyQueue<HivePartitionMetadata> partitions;
     private final Deque<Iterator<InternalHiveSplit>> fileIterators = new ConcurrentLinkedDeque<>();
-    private final boolean schedulerUsesHostAddresses;
-    private final Supplier<HoodieROTablePathFilter> hoodiePathFilterSupplier;
-    private final boolean partialAggregationsPushedDown;
+    private final PartitionLoader delegatingPartitionLoader;
 
     // Purpose of this lock:
     // * Write lock: when you need a consistent view across partitions, fileIterators, and hiveSplitSource.
@@ -156,22 +83,11 @@ public class BackgroundHiveSplitLoader
             boolean schedulerUsesHostAddresses,
             boolean partialAggregationsPushedDown)
     {
-        this.table = requireNonNull(table, "table is null");
-        this.pathDomain = requireNonNull(pathDomain, "pathDomain is null");
-        this.tableBucketInfo = requireNonNull(tableBucketInfo, "tableBucketInfo is null");
         this.loaderConcurrency = loaderConcurrency;
         checkArgument(loaderConcurrency > 0, "loaderConcurrency must be > 0, found: %s", loaderConcurrency);
-        this.session = requireNonNull(session, "session is null");
-        this.hdfsEnvironment = requireNonNull(hdfsEnvironment, "hdfsEnvironment is null");
-        this.namenodeStats = requireNonNull(namenodeStats, "namenodeStats is null");
-        this.directoryLister = requireNonNull(directoryLister, "directoryLister is null");
-        this.recursiveDirWalkerEnabled = recursiveDirWalkerEnabled;
         this.executor = requireNonNull(executor, "executor is null");
         this.partitions = new ConcurrentLazyQueue<>(requireNonNull(partitions, "partitions is null"));
-        this.hdfsContext = new HdfsContext(session, table.getDatabaseName(), table.getTableName());
-        this.schedulerUsesHostAddresses = schedulerUsesHostAddresses;
-        this.hoodiePathFilterSupplier = Suppliers.memoize(HoodieROTablePathFilter::new)::get;
-        this.partialAggregationsPushedDown = partialAggregationsPushedDown;
+        this.delegatingPartitionLoader = new DelegatingPartitionLoader(table, pathDomain, tableBucketInfo, session, hdfsEnvironment, namenodeStats, directoryLister, fileIterators, recursiveDirWalkerEnabled, schedulerUsesHostAddresses, partialAggregationsPushedDown);
     }
 
     @Override
@@ -273,7 +189,7 @@ public class BackgroundHiveSplitLoader
             if (partition == null) {
                 return COMPLETED_FUTURE;
             }
-            return loadPartition(partition);
+            return delegatingPartitionLoader.loadPartition(partition, hiveSplitSource, stopped);
         }
 
         while (splits.hasNext() && !stopped) {
@@ -286,386 +202,5 @@ public class BackgroundHiveSplitLoader
 
         // No need to put the iterator back, since it's either empty or we've stopped
         return COMPLETED_FUTURE;
-    }
-
-    private ListenableFuture<?> loadPartition(HivePartitionMetadata partition)
-            throws IOException
-    {
-        String partitionName = partition.getHivePartition().getPartitionId();
-        Storage storage = partition.getPartition().map(Partition::getStorage).orElse(table.getStorage());
-        String inputFormatName = storage.getStorageFormat().getInputFormat();
-        int partitionDataColumnCount = partition.getPartition()
-                .map(p -> p.getColumns().size())
-                .orElse(table.getDataColumns().size());
-        List<HivePartitionKey> partitionKeys = getPartitionKeys(table, partition.getPartition());
-        Path path = new Path(getPartitionLocation(table, partition.getPartition()));
-        Configuration configuration = hdfsEnvironment.getConfiguration(hdfsContext, path);
-        InputFormat<?, ?> inputFormat = getInputFormat(configuration, inputFormatName, false);
-        ExtendedFileSystem fs = hdfsEnvironment.getFileSystem(hdfsContext, path);
-        boolean s3SelectPushdownEnabled = shouldEnablePushdownForTable(session, table, path.toString(), partition.getPartition());
-
-        if (inputFormat instanceof SymlinkTextInputFormat) {
-            if (tableBucketInfo.isPresent()) {
-                throw new PrestoException(NOT_SUPPORTED, "Bucketed table in SymlinkTextInputFormat is not yet supported");
-            }
-
-            // TODO: This should use an iterator like the HiveFileIterator
-            ListenableFuture<?> lastResult = COMPLETED_FUTURE;
-            for (Path targetPath : getTargetPathsFromSymlink(fs, path)) {
-                // The input should be in TextInputFormat.
-                TextInputFormat targetInputFormat = new TextInputFormat();
-                // the splits must be generated using the file system for the target path
-                // get the configuration for the target path -- it may be a different hdfs instance
-                ExtendedFileSystem targetFilesystem = hdfsEnvironment.getFileSystem(hdfsContext, targetPath);
-                JobConf targetJob = toJobConf(targetFilesystem.getConf());
-                targetJob.setInputFormat(TextInputFormat.class);
-                targetInputFormat.configure(targetJob);
-                FileInputFormat.setInputPaths(targetJob, targetPath);
-                InputSplit[] targetSplits = targetInputFormat.getSplits(targetJob, 0);
-
-                InternalHiveSplitFactory splitFactory = new InternalHiveSplitFactory(
-                        targetFilesystem,
-                        inputFormat,
-                        pathDomain,
-                        getNodeSelectionStrategy(session),
-                        getMaxInitialSplitSize(session),
-                        s3SelectPushdownEnabled,
-                        new HiveSplitPartitionInfo(storage, path.toUri(), partitionKeys, partitionName, partitionDataColumnCount, partition.getPartitionSchemaDifference(), Optional.empty()),
-                        schedulerUsesHostAddresses,
-                        partition.getEncryptionInformation());
-                lastResult = addSplitsToSource(targetSplits, splitFactory);
-                if (stopped) {
-                    return COMPLETED_FUTURE;
-                }
-            }
-            return lastResult;
-        }
-
-        Optional<BucketConversion> bucketConversion = Optional.empty();
-        boolean bucketConversionRequiresWorkerParticipation = false;
-        if (partition.getPartition().isPresent()) {
-            Optional<HiveBucketProperty> partitionBucketProperty = partition.getPartition().get().getStorage().getBucketProperty();
-            if (tableBucketInfo.isPresent() && partitionBucketProperty.isPresent()) {
-                int tableBucketCount = tableBucketInfo.get().getTableBucketCount();
-                int partitionBucketCount = partitionBucketProperty.get().getBucketCount();
-                // Validation was done in HiveSplitManager#getPartitionMetadata.
-                // Here, it's just trying to see if its needs the BucketConversion.
-                if (tableBucketCount != partitionBucketCount) {
-                    bucketConversion = Optional.of(new BucketConversion(tableBucketCount, partitionBucketCount, tableBucketInfo.get().getBucketColumns()));
-                    if (tableBucketCount > partitionBucketCount) {
-                        bucketConversionRequiresWorkerParticipation = true;
-                    }
-                }
-            }
-        }
-        InternalHiveSplitFactory splitFactory = new InternalHiveSplitFactory(
-                fs,
-                inputFormat,
-                pathDomain,
-                getNodeSelectionStrategy(session),
-                getMaxInitialSplitSize(session),
-                s3SelectPushdownEnabled,
-                new HiveSplitPartitionInfo(
-                        storage,
-                        path.toUri(),
-                        partitionKeys,
-                        partitionName,
-                        partitionDataColumnCount,
-                        partition.getPartitionSchemaDifference(),
-                        bucketConversionRequiresWorkerParticipation ? bucketConversion : Optional.empty()),
-                schedulerUsesHostAddresses,
-                partition.getEncryptionInformation());
-
-        if (!isHudiParquetInputFormat(inputFormat) && shouldUseFileSplitsFromInputFormat(inputFormat)) {
-            if (tableBucketInfo.isPresent()) {
-                throw new PrestoException(NOT_SUPPORTED, "Presto cannot read bucketed partition in an input format with UseFileSplitsFromInputFormat annotation: " + inputFormat.getClass().getSimpleName());
-            }
-            JobConf jobConf = toJobConf(configuration);
-            FileInputFormat.setInputPaths(jobConf, path);
-            InputSplit[] splits = inputFormat.getSplits(jobConf, 0);
-
-            return addSplitsToSource(splits, splitFactory);
-        }
-        PathFilter pathFilter = isHudiParquetInputFormat(inputFormat) ? hoodiePathFilterSupplier.get() : path1 -> true;
-        // S3 Select pushdown works at the granularity of individual S3 objects,
-        // Partial aggregation pushdown works at the granularity of individual files
-        // therefore we must not split files when either is enabled.
-        Properties schema = getHiveSchema(storage.getSerdeParameters(), table.getParameters());
-        // Skip header / footer lines are not splittable except for a special case when skip.header.line.count=1
-        boolean splittable = !s3SelectPushdownEnabled && !partialAggregationsPushedDown && getFooterCount(schema) == 0 && getHeaderCount(schema) <= 1;
-
-        // Bucketed partitions are fully loaded immediately since all files must be loaded to determine the file to bucket mapping
-        if (tableBucketInfo.isPresent()) {
-            if (tableBucketInfo.get().isVirtuallyBucketed()) {
-                // For virtual bucket, bucket conversion must not be present because there is no physical partition bucket count
-                checkState(!bucketConversion.isPresent(), "Virtually bucketed table must not have partitions that are physically bucketed");
-                checkState(
-                        tableBucketInfo.get().getTableBucketCount() == tableBucketInfo.get().getReadBucketCount(),
-                        "Table and read bucket count should be the same for virtual bucket");
-                return hiveSplitSource.addToQueue(getVirtuallyBucketedSplits(path, fs, splitFactory, tableBucketInfo.get().getReadBucketCount(), splittable, pathFilter));
-            }
-            return hiveSplitSource.addToQueue(getBucketedSplits(path, fs, splitFactory, tableBucketInfo.get(), bucketConversion, partitionName, splittable, pathFilter));
-        }
-
-        fileIterators.addLast(createInternalHiveSplitIterator(path, fs, splitFactory, splittable, pathFilter, partition.getPartition()));
-        return COMPLETED_FUTURE;
-    }
-
-    private ListenableFuture<?> addSplitsToSource(InputSplit[] targetSplits, InternalHiveSplitFactory splitFactory)
-            throws IOException
-    {
-        ListenableFuture<?> lastResult = COMPLETED_FUTURE;
-        for (InputSplit inputSplit : targetSplits) {
-            Optional<InternalHiveSplit> internalHiveSplit = splitFactory.createInternalHiveSplit((FileSplit) inputSplit);
-            if (internalHiveSplit.isPresent()) {
-                lastResult = hiveSplitSource.addToQueue(internalHiveSplit.get());
-            }
-            if (stopped) {
-                return COMPLETED_FUTURE;
-            }
-        }
-        return lastResult;
-    }
-
-    private static boolean isHudiParquetInputFormat(InputFormat<?, ?> inputFormat)
-    {
-        if (inputFormat instanceof HoodieParquetRealtimeInputFormat) {
-            return false;
-        }
-        return inputFormat instanceof HoodieParquetInputFormat;
-    }
-
-    private static boolean shouldUseFileSplitsFromInputFormat(InputFormat<?, ?> inputFormat)
-    {
-        return Arrays.stream(inputFormat.getClass().getAnnotations())
-                .map(Annotation::annotationType)
-                .map(Class::getSimpleName)
-                .anyMatch(name -> name.equals("UseFileSplitsFromInputFormat"));
-    }
-
-    private Iterator<InternalHiveSplit> createInternalHiveSplitIterator(Path path, ExtendedFileSystem fileSystem, InternalHiveSplitFactory splitFactory, boolean splittable, PathFilter pathFilter, Optional<Partition> partition)
-    {
-        boolean cacheable = isUseListDirectoryCache(session);
-        if (partition.isPresent()) {
-            // Use cache only for sealed partitions
-            cacheable &= partition.get().isSealedPartition();
-        }
-
-        HiveDirectoryContext hiveDirectoryContext = new HiveDirectoryContext(recursiveDirWalkerEnabled ? RECURSE : IGNORED, cacheable);
-        return stream(directoryLister.list(fileSystem, table, path, namenodeStats, pathFilter, hiveDirectoryContext))
-                .map(status -> splitFactory.createInternalHiveSplit(status, splittable))
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .iterator();
-    }
-
-    private List<InternalHiveSplit> getBucketedSplits(
-            Path path,
-            ExtendedFileSystem fileSystem,
-            InternalHiveSplitFactory splitFactory,
-            BucketSplitInfo bucketSplitInfo,
-            Optional<BucketConversion> bucketConversion,
-            String partitionName,
-            boolean splittable,
-            PathFilter pathFilter)
-    {
-        int readBucketCount = bucketSplitInfo.getReadBucketCount();
-        int tableBucketCount = bucketSplitInfo.getTableBucketCount();
-        int partitionBucketCount = bucketConversion.map(BucketConversion::getPartitionBucketCount).orElse(tableBucketCount);
-
-        checkState(readBucketCount <= tableBucketCount, "readBucketCount(%s) should be less than or equal to tableBucketCount(%s)", readBucketCount, tableBucketCount);
-
-        // list all files in the partition
-        List<HiveFileInfo> fileInfos = new ArrayList<>(partitionBucketCount);
-        try {
-            Iterators.addAll(fileInfos, directoryLister.list(fileSystem, table, path, namenodeStats, pathFilter, new HiveDirectoryContext(FAIL, isUseListDirectoryCache(session))));
-        }
-        catch (NestedDirectoryNotAllowedException e) {
-            // Fail here to be on the safe side. This seems to be the same as what Hive does
-            throw new PrestoException(
-                    HIVE_INVALID_BUCKET_FILES,
-                    format("Hive table '%s' is corrupt. Found sub-directory in bucket directory for partition: %s",
-                            new SchemaTableName(table.getDatabaseName(), table.getTableName()),
-                            partitionName));
-        }
-
-        // verify we found one file per bucket
-        if (fileInfos.size() != partitionBucketCount) {
-            throw new PrestoException(
-                    HIVE_INVALID_BUCKET_FILES,
-                    format("Hive table '%s' is corrupt. The number of files in the directory (%s) does not match the declared bucket count (%s) for partition: %s",
-                            new SchemaTableName(table.getDatabaseName(), table.getTableName()),
-                            fileInfos.size(),
-                            partitionBucketCount,
-                            partitionName));
-        }
-
-        // Sort FileStatus objects (instead of, e.g., fileStatus.getPath().toString). This matches org.apache.hadoop.hive.ql.metadata.Table.getSortedPaths
-        fileInfos.sort(null);
-
-        // convert files internal splits
-        List<InternalHiveSplit> splitList = new ArrayList<>();
-        for (int bucketNumber = 0; bucketNumber < max(readBucketCount, partitionBucketCount); bucketNumber++) {
-            // Physical bucket #. This determine file name. It also determines the order of splits in the result.
-            int partitionBucketNumber = bucketNumber % partitionBucketCount;
-            // Logical bucket #. Each logical bucket corresponds to a "bucket" from engine's perspective.
-            int readBucketNumber = bucketNumber % readBucketCount;
-
-            boolean containsIneligibleTableBucket = false;
-            List<Integer> eligibleTableBucketNumbers = new ArrayList<>();
-            for (int tableBucketNumber = bucketNumber % tableBucketCount; tableBucketNumber < tableBucketCount; tableBucketNumber += max(readBucketCount, partitionBucketCount)) {
-                // table bucket number: this is used for evaluating "$bucket" filters.
-                if (bucketSplitInfo.isTableBucketEnabled(tableBucketNumber)) {
-                    eligibleTableBucketNumbers.add(tableBucketNumber);
-                }
-                else {
-                    containsIneligibleTableBucket = true;
-                }
-            }
-
-            if (!eligibleTableBucketNumbers.isEmpty() && containsIneligibleTableBucket) {
-                throw new PrestoException(
-                        NOT_SUPPORTED,
-                        "The bucket filter cannot be satisfied. There are restrictions on the bucket filter when all the following is true: " +
-                                "1. a table has a different buckets count as at least one of its partitions that is read in this query; " +
-                                "2. the table has a different but compatible bucket number with another table in the query; " +
-                                "3. some buckets of the table is filtered out from the query, most likely using a filter on \"$bucket\". " +
-                                "(table name: " + table.getTableName() + ", table bucket count: " + tableBucketCount + ", " +
-                                "partition bucket count: " + partitionBucketCount + ", effective reading bucket count: " + readBucketCount + ")");
-            }
-            if (!eligibleTableBucketNumbers.isEmpty()) {
-                HiveFileInfo fileInfo = fileInfos.get(partitionBucketNumber);
-                eligibleTableBucketNumbers.stream()
-                        .map(tableBucketNumber -> splitFactory.createInternalHiveSplit(fileInfo, readBucketNumber, tableBucketNumber, splittable))
-                        .forEach(optionalSplit -> optionalSplit.ifPresent(splitList::add));
-            }
-        }
-        return splitList;
-    }
-
-    private List<InternalHiveSplit> getVirtuallyBucketedSplits(Path path, ExtendedFileSystem fileSystem, InternalHiveSplitFactory splitFactory, int bucketCount, boolean splittable, PathFilter pathFilter)
-    {
-        // List all files recursively in the partition and assign virtual bucket number to each of them
-        HiveDirectoryContext hiveDirectoryContext = new HiveDirectoryContext(recursiveDirWalkerEnabled ? RECURSE : IGNORED, isUseListDirectoryCache(session));
-        return stream(directoryLister.list(fileSystem, table, path, namenodeStats, pathFilter, hiveDirectoryContext))
-                .map(fileInfo -> {
-                    int virtualBucketNumber = getVirtualBucketNumber(bucketCount, fileInfo.getPath());
-                    return splitFactory.createInternalHiveSplit(fileInfo, virtualBucketNumber, virtualBucketNumber, splittable);
-                })
-                .filter(Optional::isPresent)
-                .map(Optional::get)
-                .collect(toImmutableList());
-    }
-
-    private static List<Path> getTargetPathsFromSymlink(ExtendedFileSystem fileSystem, Path symlinkDir)
-    {
-        try {
-            FileStatus[] symlinks = fileSystem.listStatus(symlinkDir, HIDDEN_FILES_PATH_FILTER);
-            List<Path> targets = new ArrayList<>();
-
-            for (FileStatus symlink : symlinks) {
-                try (BufferedReader reader = new BufferedReader(new InputStreamReader(fileSystem.open(symlink.getPath()), StandardCharsets.UTF_8))) {
-                    CharStreams.readLines(reader).stream()
-                            .map(Path::new)
-                            .forEach(targets::add);
-                }
-            }
-            return targets;
-        }
-        catch (IOException e) {
-            throw new PrestoException(HIVE_BAD_DATA, "Error parsing symlinks from: " + symlinkDir, e);
-        }
-    }
-
-    private static List<HivePartitionKey> getPartitionKeys(Table table, Optional<Partition> partition)
-    {
-        if (!partition.isPresent()) {
-            return ImmutableList.of();
-        }
-        ImmutableList.Builder<HivePartitionKey> partitionKeys = ImmutableList.builder();
-        List<Column> keys = table.getPartitionColumns();
-        List<String> values = partition.get().getValues();
-        checkCondition(keys.size() == values.size(), HIVE_INVALID_METADATA, "Expected %s partition key values, but got %s", keys.size(), values.size());
-        for (int i = 0; i < keys.size(); i++) {
-            String name = keys.get(i).getName();
-            HiveType hiveType = keys.get(i).getType();
-            if (!hiveType.isSupportedType()) {
-                throw new PrestoException(NOT_SUPPORTED, format("Unsupported Hive type %s found in partition keys of table %s.%s", hiveType, table.getDatabaseName(), table.getTableName()));
-            }
-            String value = values.get(i);
-            checkCondition(value != null, HIVE_INVALID_PARTITION_VALUE, "partition key value cannot be null for field: %s", name);
-            partitionKeys.add(new HivePartitionKey(name, value));
-        }
-        return partitionKeys.build();
-    }
-
-    public static class BucketSplitInfo
-    {
-        private final List<HiveColumnHandle> bucketColumns;
-        private final int tableBucketCount;
-        private final int readBucketCount;
-        private final IntPredicate bucketFilter;
-
-        public static Optional<BucketSplitInfo> createBucketSplitInfo(Optional<HiveBucketHandle> bucketHandle, Optional<HiveBucketFilter> bucketFilter)
-        {
-            requireNonNull(bucketHandle, "bucketHandle is null");
-            requireNonNull(bucketFilter, "buckets is null");
-
-            if (!bucketHandle.isPresent()) {
-                checkArgument(!bucketFilter.isPresent(), "bucketHandle must be present if bucketFilter is present");
-                return Optional.empty();
-            }
-
-            int tableBucketCount = bucketHandle.get().getTableBucketCount();
-            int readBucketCount = bucketHandle.get().getReadBucketCount();
-
-            List<HiveColumnHandle> bucketColumns = bucketHandle.get().getColumns();
-            IntPredicate predicate = bucketFilter
-                    .<IntPredicate>map(filter -> filter.getBucketsToKeep()::contains)
-                    .orElse(bucket -> true);
-            return Optional.of(new BucketSplitInfo(bucketColumns, tableBucketCount, readBucketCount, predicate));
-        }
-
-        private BucketSplitInfo(List<HiveColumnHandle> bucketColumns, int tableBucketCount, int readBucketCount, IntPredicate bucketFilter)
-        {
-            this.bucketColumns = ImmutableList.copyOf(requireNonNull(bucketColumns, "bucketColumns is null"));
-            this.tableBucketCount = tableBucketCount;
-            this.readBucketCount = readBucketCount;
-            this.bucketFilter = requireNonNull(bucketFilter, "bucketFilter is null");
-        }
-
-        public List<HiveColumnHandle> getBucketColumns()
-        {
-            return bucketColumns;
-        }
-
-        public int getTableBucketCount()
-        {
-            return tableBucketCount;
-        }
-
-        public int getReadBucketCount()
-        {
-            return readBucketCount;
-        }
-
-        public boolean isVirtuallyBucketed()
-        {
-            return bucketColumns.size() == 1 && bucketColumns.get(0).equals(pathColumnHandle());
-        }
-
-        /**
-         * Evaluates whether the provided table bucket number passes the bucket predicate.
-         * A bucket predicate can be present in two cases:
-         * <ul>
-         * <li>Filter on "$bucket" column. e.g. {@code "$bucket" between 0 and 100}
-         * <li>Single-value equality filter on all bucket columns. e.g. for a table with two bucketing columns,
-         * {@code bucketCol1 = 'a' AND bucketCol2 = 123}
-         * </ul>
-         */
-        public boolean isTableBucketEnabled(int tableBucketNumber)
-        {
-            return bucketFilter.test(tableBucketNumber);
-        }
     }
 }
